src/requirements.txt
wheel
prometheus-client
pyserial
